refactor(stdlib): expose baml.math as float[] methods and drop the namespace (B-712)
Migrate the `baml.math` aggregation builtins to methods and remove the
`baml.math` namespace entirely:

- `baml.math.sum`    -> `float[].sum()`   (Summable; backed by the private root
  native `_sum_float`)
- `baml.math.mean`   -> `float[].mean()`   (new `FloatStats` interface)
- `baml.math.median` -> `float[].median()`
- `baml.math._sum_int` -> private root native `_sum_int` (still backs `int[].sum()`)
- `baml.math.trunc`  -> private root native `_trunc_to_int` (saturating, never
  throws); the three internal `baml.llm` retry-delay callers are repointed to it.
  Public code should use `float.itrunc()` / `float.trunc()`.

The native impls move from the deleted `BamlNamespaceMath` into `BamlPackageBaml`
(package_baml/root.rs). Entry-point tests that used `baml.math.trunc` now use the
public native `baml.sys.now_ms`; `_trunc_to_int`'s saturation semantics get a
direct unit test. Snapshots regenerated for the shifted global function table.
