+++ baml_language/crates/baml_builtins2/baml_std/baml/float.baml
@@ -448,3 +448,16 @@ class Float {
     $rust_function
   }
 }
+
+// ─── Internal (private) native helpers ────────────────────────────────────────
+
+/// Truncates `value` toward zero and returns the integer part, **saturating**
+/// to the `int` range and mapping NaN to `0` — it never throws.
+///
+/// Private helper (leading `_`; formerly the public `baml.math.trunc`) backing
+/// internal retry-delay math in `baml.llm`. User code should prefer the
+/// range-checked, throwing `float.itrunc()` (or the float-returning
+/// `float.trunc()`); this saturating form is intentionally not public surface.
+function _trunc_to_int(value: float) -> int throws never {
+  $rust_function
+}
