+++ baml_language/sdk_tests/crates/python_pydantic2/function_calls/customizable/test_stdlib_entrypoints.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from baml_sdk.baml.fs import exists
-from baml_sdk.baml.math import trunc
+from baml_sdk.baml.sys import now_ms
 
 
 def _generated_sdk_file(rel_path: str) -> str | None:
@@ -15,11 +15,11 @@ def _generated_sdk_file(rel_path: str) -> str | None:
     return path.read_text(encoding="utf-8")
 
 
-# `baml.math.trunc(value: float) -> int` is a `$rust_function` →
-# `FunctionKind::Native`. Calling it as an entry point should truncate toward
-# zero and return `3`, not reject with `NotInvokableAsEntry`.
-def test_native_trunc_callable_as_entry_point():
-    assert trunc(3.7) == 3
+# `baml.sys.now_ms() -> int` is a `$rust_function` → `FunctionKind::Native`.
+# Calling it as an entry point should run the native and return a positive
+# millisecond timestamp, not reject with `NotInvokableAsEntry`.
+def test_native_now_ms_callable_as_entry_point():
+    assert now_ms() > 0
 
 
 # `baml.fs.exists(path: string) -> bool` is a `$rust_io_function` →
